refactor(api): split users.controller god file into focused controllers (ATT-429)
Split the ~1220-line UsersController into five concern-focused controllers
(registration, invitations, profile/self-service, admin CRUD, permissions)
and push business logic into dedicated services (SignupDomainService,
UserRegistrationService, UserPasswordService, UserInvitationService,
UserPermissionsService) plus a shared mapEmailSendError util.

Routes, DTOs, operationIds, @ApiTags('Users') and the generated OpenAPI
output are unchanged (verified: jq-normalized swagger.json diff is empty).
Controller registration order in the module is load-bearing and preserved
so static GET/PATCH paths are not shadowed by ':id' routes; GET
'with-permission' stays intentionally shadowed to keep pre-refactor behavior.

Tests redistributed to the new services/controllers; full api suite green.
